更新demo

+++ cloud2021/spring-boot-demo/src/main/java/com/study/springboot/serivce/PaymentService.java
@@ -8,5 +8,5 @@ public interface PaymentService
 {
     public int create(Payment payment);
 
-    public Payment getPaymentById(@Param("id") Integer id);
+    public Payment getPaymentById(@Param("id") Long id);
 }
